+++ vector.c
@@ -58,7 +58,7 @@ void *vector_back(vector_t *vector)
 
 void *vector_get_data(vector_t *vector, int i)
 {
-    if (i < vector->c) {
+    if ((unsigned int) i < vector->c) {
         return vector->buf[i].data;
     }
     return NULL;
